Add legacy embedded headers format support
It is really expected in real application that data in the Kinesis stream
might be produced by the previous Kinesis binder version.
The embedded headers support was present via `EmbeddedHeaderUtils`.
A new version of the binder is based on an `EmbeddedJsonHeadersMessageMapper`
from Spring Integration.

* Introduce `LegacyEmbeddedHeadersSupportBytesMessageMapper`
to try an `EmbeddedHeaderUtils` first and then fall back to `EmbeddedJsonHeadersMessageMapper`
* Use this new class as an `embeddedHeadersMapper` in the Kinesis message-driven channel adapters
on consumer binding endpoints.